--- torchtitan/components/distributed_optimizers/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""Distributed optimizer implementations and redistribution runtimes."""
