+++ torchtitan/distributed/flex_shard/_optimizer_reshard_schedule.py
@@ -4,83 +4,27 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""Public bucket contracts and private resharding for distributed optimizers.
-
-The high-level load-balanced scheduling, packed all-to-all redistribution, and
-asynchronous optimizer-compute overlap are informed by Canzona and its FSDP
-implementation:
-
-- Canzona: https://arxiv.org/abs/2602.06079
-- FSDP-Canzona: https://github.com/liangyuwang/FSDP-Canzona
-"""
+"""Build tensor-region routes and packed collective schedules."""
 
 from __future__ import annotations
 
 import fnmatch
 import hashlib
 import math
-from collections import deque
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass, field
 from enum import Enum
-from types import ModuleType
 from typing import Any, cast, Generic, TypeAlias, TypeVar
 
 import torch
 import torch.distributed as dist
-from torch import Tensor
 from torch.distributed.device_mesh import DeviceMesh
 from torch.distributed.tensor import DTensor, Replicate, Shard
 
-
-__all__ = ["BucketConfig", "BucketSpec"]
-
-
-@dataclass(frozen=True, slots=True)
-class BucketConfig:
-    """Static bucket configuration resolved after runtime meshes exist.
-
-    ``mesh_axis`` is the storage mesh axis used for redistribution. Compute
-    distributions and any single-rank assignments come from the optimizer's
-    parameter metadata. Parameters requiring redistribution determine the
-    concrete communication mesh. An entirely local bucket does not bind a
-    communication mesh.
-    """
-
-    patterns: tuple[str, ...]
-    mesh_axis: str
-    name: str = ""
-
-    def __post_init__(self) -> None:
-        object.__setattr__(self, "patterns", tuple(self.patterns))
-
-    def bind(self, mesh: DeviceMesh | None) -> BucketSpec:
-        return BucketSpec(
-            patterns=self.patterns,
-            mesh=mesh,
-            name=self.name,
-        )
+from .optimizer_reshard import _BucketSpec, BucketConfig
 
 
-@dataclass(frozen=True, slots=True)
-class BucketSpec:
-    """One ordered optimizer-work bucket selected by canonical FQN.
-
-    Patterns use case-sensitive ``fnmatch`` syntax. Every optimizer FQN must
-    match exactly one bucket, and sequence order controls execution order.
-    ``mesh`` is the bucket's exact one-dimensional communication mesh, or
-    ``None`` when every matched parameter is already compute-ready. ``name`` is
-    diagnostic metadata only.
-    """
-
-    patterns: tuple[str, ...]
-    mesh: DeviceMesh | None
-    name: str = ""
-
-    def __post_init__(self) -> None:
-        if self.mesh is not None and self.mesh.ndim != 1:
-            raise ValueError("BucketSpec mesh must be one-dimensional")
-        object.__setattr__(self, "patterns", tuple(self.patterns))
+__all__: list[str] = []
 
 
 _ItemT = TypeVar("_ItemT")
@@ -94,7 +38,7 @@ def _bind_bucket_configs(
     get_storage_dtensor: Callable[[_ItemT], DTensor],
     requires_redistribution: Callable[[_ItemT], bool],
     get_required_storage_mesh_axis: Callable[[_ItemT], int | None],
-) -> tuple[BucketSpec, ...]:
+) -> tuple[_BucketSpec, ...]:
     """Bind configs after each item's redistribution requirement is resolved.
 
     Redistributed items determine the communication mesh. An entirely local
@@ -116,7 +60,7 @@ def _bind_bucket_configs(
         if not matched_items:
             continue
         if not redistributed_items:
-            specs.append(config.bind(None))
+            specs.append(config._bind(None))
             continue
 
         mesh_axis = config.mesh_axis
@@ -148,13 +92,13 @@ def _bind_bucket_configs(
             raise ValueError(
                 f"bucket {config.name!r} resolves to inconsistent communication meshes"
             )
-        specs.append(config.bind(mesh))
+        specs.append(config._bind(mesh))
     return tuple(specs)
 
 
 def _resolve_buckets(
     items: Sequence[_ItemT],
-    specs: Sequence[BucketSpec],
+    specs: Sequence[_BucketSpec],
     *,
     get_fqn: Callable[[_ItemT], str],
 ) -> tuple[tuple[_ItemT, ...], ...]:
@@ -609,6 +553,73 @@ def _build_single_participant_redistribution_plan(
     )
 
 
+def _build_replicated_to_dim0_shard_plan(
+    storage_regions: Sequence[tuple[tuple[int, ...], _TensorRegion]],
+    *,
+    participants: tuple[int, ...],
+    logical_shape: tuple[int, ...],
+) -> _RedistributionPlan:
+    """Partition replicated storage along compute dimension 0."""
+    full_region = _TensorRegion(
+        offsets=(0,) * len(logical_shape),
+        shape=logical_shape,
+    )
+    _require_valid_plan(
+        tuple(storage_regions) == ((participants, full_region),),
+        "dim-0 sharded compute requires replicated storage",
+    )
+
+    storage_partitions = tuple(
+        _ParticipantPartition(
+            participant=participant,
+            tensor_shape=logical_shape,
+            logical_regions=(full_region,),
+        )
+        for participant in participants
+    )
+    compute_partitions = []
+    storage_to_compute_routes = []
+    for participant_index, participant in enumerate(participants):
+        local_dim0, dim0_offset = Shard.local_shard_size_and_offset(
+            logical_shape[0],
+            len(participants),
+            participant_index,
+        )
+        local_shape = (local_dim0, *logical_shape[1:])
+        logical_region = _TensorRegion(
+            offsets=(dim0_offset,) + (0,) * (len(logical_shape) - 1),
+            shape=local_shape,
+        )
+        tensor_region = _TensorRegion(
+            offsets=(0,) * len(logical_shape),
+            shape=local_shape,
+        )
+        compute_partitions.append(
+            _ParticipantPartition(
+                participant=participant,
+                tensor_shape=local_shape,
+                logical_regions=(logical_region,),
+            )
+        )
+        if not logical_region.numel:
+            continue
+        storage_to_compute_routes.append(
+            _TensorRegionRoute(
+                logical_region=logical_region,
+                source=_RouteEndpoint(logical_region, participants),
+                destination=_RouteEndpoint(tensor_region, (participant,)),
+            )
+        )
+
+    return _RedistributionPlan(
+        participants=participants,
+        logical_shape=logical_shape,
+        storage_partitions=storage_partitions,
+        compute_partitions=tuple(compute_partitions),
+        storage_to_compute_routes=tuple(storage_to_compute_routes),
+    )
+
+
 @dataclass(frozen=True, slots=True)
 class _PackedSpan:
     """Physical packed-buffer location for an endpoint tensor region."""
@@ -642,18 +653,6 @@ class _PackedAllToAllSchedule:
             and self.input_buffer_numel != self.output_buffer_numel
         ):
             raise RuntimeError("local all-to-all schedule has mismatched buffers")
-
-    def execute(self, output: Tensor, input: Tensor) -> None:
-        if not self.has_remote_transfers:
-            output[: self.output_buffer_numel].copy_(input[: self.input_buffer_numel])
-            return
-        dist.all_to_all_single(
-            output[: self.output_buffer_numel],
-            input[: self.input_buffer_numel],
-            output_split_sizes=self.output_split_sizes,
-            input_split_sizes=self.input_split_sizes,
-            group=self.process_group,
-        )
 
 
 @dataclass(frozen=True, slots=True)
@@ -694,719 +693,6 @@ _BucketPlan: TypeAlias = _LocalBucketPlan[_ItemT] | _RedistributionBucketPlan[_I
 class _BucketPlanningResult(Generic[_ItemT]):
     plans: tuple[_BucketPlan[_ItemT], ...]
     ordered_items: tuple[_ItemT, ...]
-
-
-_BufferKey: TypeAlias = tuple[torch.device, torch.dtype]
-
-
-@dataclass(slots=True)
-class _BufferRequirements:
-    storage_exchange_numel: int | None = None
-    compute_exchange_numel: int | None = None
-    compute_scratch_numel: int | None = None
-    storage_scratch_numel: int | None = None
-
-    def include_communication(
-        self,
-        *,
-        storage_exchange_numel: int,
-        compute_exchange_numel: int,
-    ) -> None:
-        self.storage_exchange_numel = max(
-            self.storage_exchange_numel or 0,
-            storage_exchange_numel,
-        )
-        self.compute_exchange_numel = max(
-            self.compute_exchange_numel or 0,
-            compute_exchange_numel,
-        )
-
-    def include_compute_scratch(self, numel: int) -> None:
-        self.compute_scratch_numel = max(self.compute_scratch_numel or 0, numel)
-
-    def include_storage_scratch(self, numel: int) -> None:
-        self.storage_scratch_numel = max(self.storage_scratch_numel or 0, numel)
-
-
-@dataclass(slots=True)
-class _ReservedBuffers:
-    storage_exchange: Tensor | None = None
-    compute_exchange: Tensor | None = None
-    # These remain separate because transfer-stream preparation/finalization
-    # can overlap caller-stream compute within the same pipeline slot.
-    compute_scratch: Tensor | None = None
-    storage_scratch: Tensor | None = None
-
-
-def _reserve_tensor(
-    tensor: Tensor | None,
-    *,
-    numel: int,
-    dtype: torch.dtype,
-    device: torch.device,
-) -> Tensor:
-    if tensor is None or tensor.numel() < numel:
-        return torch.empty(numel, dtype=dtype, device=device)
-    return tensor
-
-
-def _reserved_view(tensor: Tensor | None, numel: int) -> Tensor:
-    assert tensor is not None and tensor.numel() >= numel
-    return tensor[:numel]
-
-
-@dataclass(slots=True)
-class _BufferSlot:
-    buffers: dict[_BufferKey, _ReservedBuffers] = field(default_factory=dict)
-    _recorded_compute_streams: set[torch.Stream] = field(default_factory=set)
-
-    def reserve(
-        self,
-        requirements: dict[_BufferKey, _BufferRequirements],
-        *,
-        device_handle: ModuleType,
-        compute_stream: torch.Stream,
-        transfer_stream: torch.Stream | None,
-    ) -> None:
-        if transfer_stream is not None:
-            with device_handle.stream(transfer_stream):
-                for (device, dtype), required in requirements.items():
-                    reserved = self.buffers.setdefault(
-                        (device, dtype), _ReservedBuffers()
-                    )
-                    if required.storage_exchange_numel is not None:
-                        reserved.storage_exchange = _reserve_tensor(
-                            reserved.storage_exchange,
-                            numel=required.storage_exchange_numel,
-                            dtype=dtype,
-                            device=device,
-                        )
-                    if required.compute_exchange_numel is not None:
-                        reserved.compute_exchange = _reserve_tensor(
-                            reserved.compute_exchange,
-                            numel=required.compute_exchange_numel,
-                            dtype=dtype,
-                            device=device,
-                        )
-                    if required.storage_scratch_numel is not None:
-                        reserved.storage_scratch = _reserve_tensor(
-                            reserved.storage_scratch,
-                            numel=required.storage_scratch_numel,
-                            dtype=dtype,
-                            device=device,
-                        )
-
-        with device_handle.stream(compute_stream):
-            for (device, dtype), required in requirements.items():
-                if required.compute_scratch_numel is None:
-                    continue
-                reserved = self.buffers.setdefault((device, dtype), _ReservedBuffers())
-                reserved.compute_scratch = _reserve_tensor(
-                    reserved.compute_scratch,
-                    numel=required.compute_scratch_numel,
-                    dtype=dtype,
-                    device=device,
-                )
-        self._recorded_compute_streams.clear()
-        self.record_compute_stream(compute_stream)
-
-    def record_compute_stream(self, stream: torch.Stream) -> None:
-        # Pipeline exchange buffers stay alive and use explicit slot events.
-        # Compute scratch also serves local-only work, so record each distinct
-        # caller for safe destruction or growth after queued compute.
-        if stream in self._recorded_compute_streams:
-            return
-        for reserved in self.buffers.values():
-            compute_scratch = reserved.compute_scratch
-            if compute_scratch is not None and compute_scratch.device.type != "cpu":
-                compute_scratch.record_stream(stream)
-        self._recorded_compute_streams.add(stream)
-
-    def communication_buffers(
-        self, plan: _RedistributionBucketPlan[Any]
-    ) -> tuple[Tensor, Tensor]:
-        to_compute = plan.storage_to_compute_schedule
-        to_storage = plan.compute_to_storage_schedule
-        reserved = self.buffers[(plan.device, plan.dtype)]
-        return (
-            _reserved_view(
-                reserved.storage_exchange,
-                max(
-                    to_compute.input_buffer_numel,
-                    to_storage.output_buffer_numel,
-                ),
-            ),
-            _reserved_view(
-                reserved.compute_exchange,
-                max(
-                    to_compute.output_buffer_numel,
-                    to_storage.input_buffer_numel,
-                ),
-            ),
-        )
-
-    def compute_buffer(
-        self,
-        shape: torch.Size | tuple[int, ...],
-        *,
-        dtype: torch.dtype,
-        device: torch.device,
-    ) -> Tensor:
-        reserved = self.buffers[(device, dtype)]
-        return _reserved_view(reserved.compute_scratch, math.prod(shape)).view(shape)
-
-    def storage_partition_buffer(
-        self,
-        shape: torch.Size | tuple[int, ...],
-        *,
-        dtype: torch.dtype,
-        device: torch.device,
-    ) -> Tensor:
-        reserved = self.buffers[(device, dtype)]
-        return _reserved_view(reserved.storage_scratch, math.prod(shape)).view(shape)
-
-
-@dataclass(slots=True)
-class _PipelineSlot:
-    """Persistent buffer keepalive and reusable cross-stream handoff events."""
-
-    buffers: _BufferSlot
-    # Transfer producer -> compute consumer.
-    compute_input_ready: torch.Event
-    # Compute consumer -> transfer return/reuse.
-    compute_done: torch.Event
-    # Transfer finalization -> subsequent caller work.
-    done: torch.Event
-
-
-@dataclass(slots=True)
-class _BucketWork(Generic[_ItemT]):
-    plan: _RedistributionBucketPlan[_ItemT]
-    slot: _PipelineSlot
-    storage_buffer: Tensor
-    compute_fragment_buffer: Tensor
-
-
-@dataclass(slots=True)
-class _CommunicationContext:
-    device_handle: ModuleType
-    transfer_stream: torch.Stream
-    slots: tuple[_PipelineSlot, ...]
-
-    @classmethod
-    def create(
-        cls,
-        device: torch.device,
-        num_pipeline_slots: int,
-    ) -> _CommunicationContext:
-        device_handle = torch.get_device_module(device)
-        transfer_stream = device_handle.Stream(device=device, priority=0)
-
-        def create_slot() -> _PipelineSlot:
-            return _PipelineSlot(
-                buffers=_BufferSlot(),
-                compute_input_ready=device_handle.Event(),
-                compute_done=device_handle.Event(),
-                done=device_handle.Event(),
-            )
-
-        return cls(
-            device_handle=device_handle,
-            transfer_stream=transfer_stream,
-            slots=tuple(create_slot() for _ in range(num_pipeline_slots)),
-        )
-
-
-def _include_compute_scratch_requirement(
-    requirements: dict[_BufferKey, _BufferRequirements],
-    item: _ItemT,
-    local_tensor_spec: Callable[[_ItemT], tuple[torch.Size, torch.dtype, torch.device]],
-) -> None:
-    shape, dtype, device = local_tensor_spec(item)
-    requirements.setdefault(
-        (device, dtype), _BufferRequirements()
-    ).include_compute_scratch(math.prod(shape))
-
-
-class _BucketedRedistributionRuntime(Generic[_ItemT]):
-    """Execute bucket plans with a rolling communication prefetch window.
-
-    ``reserve_buffers`` must be called after planning and after any replan. It
-    eagerly grows ``num_pipeline_slots`` rolling buffer slots and creates
-    their stream events, so ``run`` only reuses reserved communication
-    resources. One slot holds the current bucket and the remaining slots
-    prefetch later redistributed buckets.
-
-    ``prepare`` writes a Muon input into runtime-owned scratch, ``compute``
-    updates its runtime-owned input in place, and ``finalize`` consumes a
-    runtime-owned result before reuse. Callbacks run under the stream selected
-    by the runtime and must not retain tensors, synchronize, or call
-    ``Tensor.record_stream()``. Local-only buckets are prefetch barriers, so
-    no later redistributed ``prepare`` runs before an intervening local bucket.
-
-    Any exception is fatal: parameters or optimizer state may already be
-    updated and communication may be in flight, so callers must not reuse this
-    runtime or optimizer.
-    """
-
-    def __init__(
-        self,
-        device: torch.device,
-        *,
-        num_pipeline_slots: int = 2,
-    ) -> None:
-        assert type(num_pipeline_slots) is int and num_pipeline_slots > 0
-        # pyrefly: ignore [read-only]
-        self._device = device
-        self._num_pipeline_slots = num_pipeline_slots
-        self._context: _CommunicationContext | None = None
-        self._local_slot = _BufferSlot()
-
-    def reserve_buffers(
-        self,
-        plans: Sequence[_BucketPlan[_ItemT]],
-        *,
-        local_tensor_spec: Callable[
-            [_ItemT], tuple[torch.Size, torch.dtype, torch.device]
-        ],
-    ) -> None:
-        """Eagerly grow runtime-owned tensors to the plans' maximum sizes."""
-        context = self._context
-        created_context = False
-        if context is None and any(
-            isinstance(plan, _RedistributionBucketPlan) for plan in plans
-        ):
-            context = _CommunicationContext.create(
-                self._device,
-                self._num_pipeline_slots,
-            )
-            created_context = True
-
-        device_handle = (
-            context.device_handle
-            if context is not None
-            else torch.get_device_module(self._device)
-        )
-        compute_stream = device_handle.current_stream(self._device)
-        local_requirements: dict[_BufferKey, _BufferRequirements] = {}
-        slot_requirements = tuple({} for _ in context.slots) if context else ()
-        redistributed_index = 0
-        for plan in plans:
-            if isinstance(plan, _LocalBucketPlan):
-                for item in plan.items:
-                    _include_compute_scratch_requirement(
-                        local_requirements,
-                        item,
-                        local_tensor_spec,
-                    )
-                continue
-
-            assert context is not None
-            requirements = slot_requirements[
-                redistributed_index % len(slot_requirements)
-            ]
-            redistributed_index += 1
-            to_compute = plan.storage_to_compute_schedule
-            to_storage = plan.compute_to_storage_schedule
-            plan_requirements = requirements.setdefault(
-                (plan.device, plan.dtype), _BufferRequirements()
-            )
-            plan_requirements.include_communication(
-                storage_exchange_numel=max(
-                    to_compute.input_buffer_numel,
-                    to_storage.output_buffer_numel,
-                ),
-                compute_exchange_numel=max(
-                    to_compute.output_buffer_numel,
-                    to_storage.input_buffer_numel,
-                ),
-            )
-            participant = plan.group.local_participant
-            for redistribution_plan in plan.redistribution_plans:
-                storage_partition = redistribution_plan.storage_partition(participant)
-                plan_requirements.include_storage_scratch(
-                    math.prod(storage_partition.tensor_shape)
-                )
-                compute_partition = redistribution_plan.compute_partition(participant)
-                if compute_numel := math.prod(compute_partition.tensor_shape):
-                    plan_requirements.include_compute_scratch(compute_numel)
-            for item in plan.unredistributed_items:
-                _include_compute_scratch_requirement(
-                    requirements,
-                    item,
-                    local_tensor_spec,
-                )
-
-        self._local_slot.reserve(
-            local_requirements,
-            device_handle=device_handle,
-            compute_stream=compute_stream,
-            transfer_stream=None,
-        )
-        if context is not None:
-            for slot, requirements in zip(
-                context.slots, slot_requirements, strict=True
-            ):
-                slot.buffers.reserve(
-                    requirements,
-                    device_handle=device_handle,
-                    compute_stream=compute_stream,
-                    transfer_stream=context.transfer_stream,
-                )
-            if created_context:
-                for slot in context.slots:
-                    slot.compute_input_ready.record(context.transfer_stream)
-                    slot.compute_done.record(compute_stream)
-                    slot.done.record(context.transfer_stream)
-            self._context = context
-
-    def run(
-        self,
-        plans: Sequence[_BucketPlan[_ItemT]],
-        *,
-        local_tensor_spec: Callable[
-            [_ItemT], tuple[torch.Size, torch.dtype, torch.device]
-        ],
-        prepare: Callable[[_ItemT, Tensor], None],
-        compute: Callable[[_ItemT, Tensor], None],
-        finalize: Callable[[_ItemT, Tensor], None],
-    ) -> None:
-        context = self._context
-        handle = (
-            context.device_handle
-            if context is not None
-            else torch.get_device_module(self._device)
-        )
-        caller = handle.current_stream(self._device)
-        self._local_slot.record_compute_stream(caller)
-        if context is not None:
-            for slot in context.slots:
-                slot.buffers.record_compute_stream(caller)
-
-        transfer_stream: torch.Stream | None = None
-        completed = False
-        previous: _BucketWork[_ItemT] | None = None
-        prefetched: deque[_BucketWork[_ItemT]] = deque()
-        next_plan_to_prefetch = 0
-        redistributed_index = 0
-
-        def prefetch_next(slot: _PipelineSlot) -> _BucketWork[_ItemT] | None:
-            nonlocal next_plan_to_prefetch
-            if next_plan_to_prefetch >= len(plans):
-                return None
-            next_plan = plans[next_plan_to_prefetch]
-            if isinstance(next_plan, _LocalBucketPlan):
-                return None
-            next_plan_to_prefetch += 1
-            assert context is not None and transfer_stream is not None
-            return self._enqueue_storage_to_compute(
-                next_plan,
-                slot,
-                context,
-                prepare=prepare,
-            )
-
-        try:
-            for plan_index, plan in enumerate(plans):
-                if isinstance(plan, _LocalBucketPlan):
-                    assert not prefetched
-                    with handle.stream(caller):
-                        self._compute_without_redistribution(
-                            plan.items,
-                            self._local_slot,
-                            local_tensor_spec=local_tensor_spec,
-                            prepare=prepare,
-                            compute=compute,
-                            finalize=finalize,
-                        )
-                    next_plan_to_prefetch = plan_index + 1
-                    continue
-
-                assert context is not None
-                if transfer_stream is None:
-                    context.transfer_stream.wait_stream(caller)
-                    transfer_stream = context.transfer_stream
-                previous_work = previous
-                if prefetched:
-                    assert previous_work is not None
-                    work = prefetched.popleft()
-                    assert work.plan is plan
-                    self._enqueue_compute_to_storage(
-                        previous_work, context, finalize=finalize
-                    )
-                else:
-                    slot = context.slots[redistributed_index % len(context.slots)]
-                    if previous_work is not None and slot is previous_work.slot:
-                        self._enqueue_compute_to_storage(
-                            previous_work, context, finalize=finalize
-                        )
-                        work = self._enqueue_storage_to_compute(
-                            plan,
-                            slot,
-                            context,
-                            prepare=prepare,
-                        )
-                    else:
-                        work = self._enqueue_storage_to_compute(
-                            plan,
-                            slot,
-                            context,
-                            prepare=prepare,
-                        )
-                        if previous_work is not None:
-                            self._enqueue_compute_to_storage(
-                                previous_work, context, finalize=finalize
-                            )
-                    next_plan_to_prefetch = plan_index + 1
-
-                self._compute_bucket(
-                    work,
-                    work.slot,
-                    caller,
-                    context,
-                    local_tensor_spec=local_tensor_spec,
-                    prepare=prepare,
-                    compute=compute,
-                    finalize=finalize,
-                )
-                if previous_work is not None:
-                    self._release(previous_work, caller)
-                previous = work
-                redistributed_index += 1
-
-                while len(prefetched) < len(context.slots) - 1:
-                    slot = context.slots[
-                        (redistributed_index + len(prefetched)) % len(context.slots)
-                    ]
-                    if future_work := prefetch_next(slot):
-                        prefetched.append(future_work)
-                    else:
-                        break
-
-            assert not prefetched
-            if previous is not None:
-                assert context is not None
-                self._enqueue_compute_to_storage(previous, context, finalize=finalize)
-                self._release(previous, caller)
-            completed = True
-        finally:
-            if not completed and transfer_stream is not None:
-                # This fatal-only cleanup intentionally has no isolated mock-stream
-                # test to keep the unit-test footprint minimal. An exception makes
-                # the optimizer and training job unrecoverable.
-                # Preserve allocator lifetime ordering for work already queued
-                # on either stream without suppressing the active exception.
-                transfer_stream.wait_stream(caller)
-                caller.wait_stream(transfer_stream)
-
-    @staticmethod
-    def _enqueue_storage_to_compute(
-        plan: _RedistributionBucketPlan[_ItemT],
-        slot: _PipelineSlot,
-        context: _CommunicationContext,
-        *,
-        prepare: Callable[[_ItemT, Tensor], None],
-    ) -> _BucketWork[_ItemT]:
-        handle = context.device_handle
-        transfer = context.transfer_stream
-        with handle.stream(transfer):
-            (
-                storage_buffer,
-                compute_fragment_buffer,
-            ) = slot.buffers.communication_buffers(plan)
-            work = _BucketWork(plan, slot, storage_buffer, compute_fragment_buffer)
-            _prepare_redistributed(
-                plan,
-                slot.buffers,
-                storage_buffer,
-                prepare=prepare,
-            )
-            plan.storage_to_compute_schedule.execute(
-                output=compute_fragment_buffer,
-                input=storage_buffer,
-            )
-            slot.compute_input_ready.record(transfer)
-        return work
-
-    @staticmethod
-    def _compute_bucket(
-        work: _BucketWork[_ItemT],
-        slot: _PipelineSlot,
-        caller_stream: torch.Stream,
-        context: _CommunicationContext,
-        *,
-        local_tensor_spec: Callable[
-            [_ItemT], tuple[torch.Size, torch.dtype, torch.device]
-        ],
-        prepare: Callable[[_ItemT, Tensor], None],
-        compute: Callable[[_ItemT, Tensor], None],
-        finalize: Callable[[_ItemT, Tensor], None],
-    ) -> None:
-        handle = context.device_handle
-        with handle.stream(caller_stream):
-            _BucketedRedistributionRuntime._compute_without_redistribution(
-                work.plan.unredistributed_items,
-                slot.buffers,
-                local_tensor_spec=local_tensor_spec,
-                prepare=prepare,
-                compute=compute,
-                finalize=finalize,
-            )
-            caller_stream.wait_event(slot.compute_input_ready)
-            _compute_redistributed(
-                work,
-                slot.buffers,
-                compute=compute,
-            )
-            slot.compute_done.record(caller_stream)
-
-    @staticmethod
-    def _enqueue_compute_to_storage(
-        work: _BucketWork[_ItemT],
-        context: _CommunicationContext,
-        *,
-        finalize: Callable[[_ItemT, Tensor], None],
-    ) -> None:
-        handle = context.device_handle
-        transfer = context.transfer_stream
-        with handle.stream(transfer):
-            transfer.wait_event(work.slot.compute_done)
-            work.plan.compute_to_storage_schedule.execute(
-                output=work.storage_buffer,
-                input=work.compute_fragment_buffer,
-            )
-            _finalize_redistributed(work, work.slot.buffers, finalize=finalize)
-            work.slot.done.record(transfer)
-
-    @staticmethod
-    def _release(work: _BucketWork[_ItemT], caller_stream: torch.Stream) -> None:
-        caller_stream.wait_event(work.slot.done)
-
-    @staticmethod
-    def _compute_without_redistribution(
-        items: Sequence[_ItemT],
-        slot: _BufferSlot,
-        *,
-        local_tensor_spec: Callable[
-            [_ItemT], tuple[torch.Size, torch.dtype, torch.device]
-        ],
-        prepare: Callable[[_ItemT, Tensor], None],
-        compute: Callable[[_ItemT, Tensor], None],
-        finalize: Callable[[_ItemT, Tensor], None],
-    ) -> None:
-        for item in items:
-            shape, dtype, device = local_tensor_spec(item)
-            prepared = slot.compute_buffer(shape, dtype=dtype, device=device)
-            prepare(item, prepared)
-            compute(item, prepared)
-            finalize(item, prepared)
-
-
-def _prepare_redistributed(
-    plan: _RedistributionBucketPlan[_ItemT],
-    slot: _BufferSlot,
-    storage_buffer: Tensor,
-    *,
-    prepare: Callable[[_ItemT, Tensor], None],
-) -> None:
-    schedule = plan.storage_to_compute_schedule
-    participant = plan.group.local_participant
-    for index, (item, redistribution_plan) in enumerate(
-        zip(
-            plan.redistributed_items,
-            plan.redistribution_plans,
-            strict=True,
-        )
-    ):
-        partition = redistribution_plan.storage_partition(participant)
-        prepared = slot.storage_partition_buffer(
-            partition.tensor_shape,
-            dtype=plan.dtype,
-            device=plan.device,
-        )
-        prepare(item, prepared)
-        spans = schedule.input_spans_by_parameter[index]
-        for span in spans:
-            packed = storage_buffer[
-                span.buffer_offset : span.buffer_offset + span.numel
-            ]
-            packed.copy_(_tensor_region_view(prepared, span.region).reshape(-1))
-
-
-def _compute_redistributed(
-    work: _BucketWork[_ItemT],
-    slot: _BufferSlot,
-    *,
-    compute: Callable[[_ItemT, Tensor], None],
-) -> None:
-    plan = work.plan
-    participant = plan.group.local_participant
-    to_compute = plan.storage_to_compute_schedule
-    to_storage = plan.compute_to_storage_schedule
-    for index, (item, redistribution_plan) in enumerate(
-        zip(
-            plan.redistributed_items,
-            plan.redistribution_plans,
-            strict=True,
-        )
-    ):
-        partition = redistribution_plan.compute_partition(participant)
-        if not math.prod(partition.tensor_shape):
-            continue
-        received_spans = to_compute.output_spans_by_parameter[index]
-        compute_tensor = slot.compute_buffer(
-            partition.tensor_shape,
-            dtype=plan.dtype,
-            device=plan.device,
-        )
-        for span in received_spans:
-            received = work.compute_fragment_buffer[
-                span.buffer_offset : span.buffer_offset + span.numel
-            ]
-            _tensor_region_view(compute_tensor, span.region).copy_(
-                received.view(span.region.shape)
-            )
-
-        compute(item, compute_tensor)
-
-        for span in to_storage.input_spans_by_parameter[index]:
-            packed = work.compute_fragment_buffer[
-                span.buffer_offset : span.buffer_offset + span.numel
-            ]
-            packed.copy_(_tensor_region_view(compute_tensor, span.region).reshape(-1))
-
-
-def _finalize_redistributed(
-    work: _BucketWork[_ItemT],
-    slot: _BufferSlot,
-    *,
-    finalize: Callable[[_ItemT, Tensor], None],
-) -> None:
-    plan = work.plan
-    participant = plan.group.local_participant
-    schedule = work.plan.compute_to_storage_schedule
-    for index, (item, redistribution_plan) in enumerate(
-        zip(
-            plan.redistributed_items,
-            plan.redistribution_plans,
-            strict=True,
-        )
-    ):
-        partition = redistribution_plan.storage_partition(participant)
-        update = slot.storage_partition_buffer(
-            partition.tensor_shape,
-            dtype=plan.dtype,
-            device=plan.device,
-        )
-        spans = schedule.output_spans_by_parameter[index]
-        for span in spans:
-            packed = work.storage_buffer[
-                span.buffer_offset : span.buffer_offset + span.numel
-            ]
-            _tensor_region_view(update, span.region).copy_(
-                packed.view(span.region.shape)
-            )
-        finalize(item, update)
 
 
 def _resolve_routes_to_transfers(
@@ -1646,7 +932,7 @@ def _require_valid_planner_result(condition: bool, message: str) -> None:
 
 def _build_bucket_plans(
     items: Sequence[_ItemT],
-    specs: Sequence[BucketSpec],
+    specs: Sequence[_BucketSpec],
     *,
     get_fqn: Callable[[_ItemT], str],
     get_storage_dtensor: Callable[[_ItemT], DTensor],
@@ -1804,7 +1090,6 @@ def _validate_bucket_plans_across_ranks(
     plans: Sequence[_BucketPlan[_ItemT]],
     *,
     item_signature: Callable[[_ItemT], tuple[Any, ...]],
-    runtime_signature: tuple[Any, ...],
 ) -> None:
     """Collectively verify rank-stable redistribution plans.
 
@@ -1817,7 +1102,6 @@ def _validate_bucket_plans_across_ranks(
         if isinstance(plan, _LocalBucketPlan):
             continue
         description = (
-            runtime_signature,
             str(plan.dtype),
             plan.device.type,
             tuple(
@@ -1840,16 +1124,6 @@ def _validate_bucket_plans_across_ranks(
         dist.all_gather(gathered, local_hash, group=process_group)
         if any(value.item() != plan_hash for value in gathered):
             raise RuntimeError("optimizer bucket plans differ across ranks")
-
-
-def _tensor_region_view(tensor: Tensor, region: _TensorRegion) -> Tensor:
-    view = tensor[
-        tuple(
-            slice(offset, offset + size)
-            for offset, size in zip(region.offsets, region.shape, strict=True)
-        )
-    ]
-    return view
 
 
 def _redistribution_plan_key(plan: _RedistributionPlan) -> tuple[Any, ...]:
