+++ torchtitan/models/kimi_k2_7/README.md
@@ -61,6 +61,10 @@ Test scripts:
 
 ## TODO
 
+- Add DistributedMuon support for TP-produced `_StridedShard` parameter
+  layouts; tracked in [#3353](https://github.com/pytorch/torchtitan/issues/3353).
+- Add DistributedMuon support for PP stage-local optimizer groups; follow-up
+  [#4102](https://github.com/pytorch/torchtitan/pull/4102).
 - Add a video dataset training pipeline.
 - Add INT4 (compressed-tensors) checkpoint loading. The released K2.5, K2.6,
   and K2.7-Code 1T checkpoints are INT4 group-quantized; the inherited
